Conflicts:
	src/global.h
	src/global.h.in
	src/mapper_config.h.in
	symbol sets/10000/ISOM_10000.omap
	symbol sets/15000/ISOM_15000.omap
	symbol sets/4000/ISSOM_4000.omap
	symbol sets/5000/ISSOM_5000.omap

+++ src/template_dock_widget.cpp
@@ -20,9 +20,13 @@
 
 #include "template_dock_widget.h"
 
-#include <assert.h>
+#include <cassert>
 
+#if QT_VERSION < 0x050000
 #include <QtGui>
+#else
+#include <QtWidgets>
+#endif
 
 #include "map.h"
 #include "template.h"
@@ -30,6 +34,8 @@
 #include "template_adjust.h"
 #include "template_tool_move.h"
 #include "template_position_dock_widget.h"
+#include "georeferencing.h"
+#include "settings.h"
 
 /** Parses a user-entered opacity value. Values must be strings of the form "F%" where F is any decimal number between 0 and
  *  100 (inclusive), or "F", where F is any floating-point number between 0.0 and 1.0 (inclusive). Leading and trailing
@@ -94,7 +100,9 @@ TemplateWidget::TemplateWidget(Map* map, MapView* main_view, MapEditorController
 	new_button->setEnabled(false);	// TODO!
 	
 	QPushButton* open_button = new QPushButton(QIcon(":/images/open.png"), tr("Open..."));
-	delete_button = new QPushButton(QIcon(":/images/minus.png"), tr("Delete"));
+	delete_button = new QPushButton(QIcon(":/images/minus.png"), "");
+	updateDeleteButtonText();
+	connect(&Settings::getInstance(), SIGNAL(settingsChanged()), this, SLOT(updateDeleteButtonText()));
 	duplicate_button = new QPushButton(QIcon(":/images/copy.png"), tr("Duplicate"));
 	move_up_button = new QPushButton(QIcon(":/images/arrow-up.png"), tr("Move Up"));
 	move_down_button = new QPushButton(QIcon(":/images/arrow-down.png"), tr("Move Down"));
@@ -113,8 +121,10 @@ TemplateWidget::TemplateWidget(Map* map, MapView* main_view, MapEditorController
 	list_buttons_group->setLayout(list_buttons_group_layout);
 	
 	// Active group
-	active_buttons_group = new QGroupBox(tr("Selected template(s)"));
+	active_buttons_group = new QGroupBox();
 	
+	georef_button = new QPushButton();
+	georef_button->setCheckable(true);
 	move_by_hand_action = new QAction(QIcon(":/images/move.png"), tr("Move by hand"), this);
 	move_by_hand_action->setCheckable(true);
 	move_by_hand_button = new QToolButton();
@@ -127,7 +137,7 @@ TemplateWidget::TemplateWidget(Map* map, MapView* main_view, MapEditorController
 	position_button = new QPushButton(tr("Positioning..."));
 	position_button->setCheckable(true);
 	
-	more_button = new QToolButton();
+	/*more_button = new QToolButton();
 	more_button->setText(tr("More..."));
 	more_button->setPopupMode(QToolButton::InstantPopup);
 	more_button->setSizePolicy(QSizePolicy(QSizePolicy::Minimum, QSizePolicy::Fixed));
@@ -135,15 +145,17 @@ TemplateWidget::TemplateWidget(Map* map, MapView* main_view, MapEditorController
 	more_button_menu->addAction(QIcon(":/images/window-new.png"), tr("Numeric transformation window"));
 	more_button_menu->addAction(tr("Set transparent color..."));
 	more_button_menu->addAction(tr("Trace lines..."));
-	more_button->setMenu(more_button_menu);
+	more_button->setMenu(more_button_menu);*/
 	
+	int row = 0;
 	QGridLayout* active_buttons_group_layout = new QGridLayout();
 	active_buttons_group_layout->setMargin(0);
-	active_buttons_group_layout->addWidget(move_by_hand_button, 0, 0);
-	active_buttons_group_layout->addWidget(adjust_button, 0, 1);
-	//active_buttons_group_layout->addWidget(group_button, 1, 0);
-	active_buttons_group_layout->addWidget(position_button, 1, 0);
-	active_buttons_group_layout->addWidget(more_button, 1, 1);
+	active_buttons_group_layout->addWidget(georef_button, row, 0);
+	active_buttons_group_layout->addWidget(adjust_button, row++, 1);
+	active_buttons_group_layout->addWidget(move_by_hand_button, row, 0);
+	active_buttons_group_layout->addWidget(position_button, row++, 1);
+	//active_buttons_group_layout->addWidget(group_button, row++, 0);
+	//active_buttons_group_layout->addWidget(more_button, row++, 1);
 	active_buttons_group->setLayout(active_buttons_group_layout);
 	
 	selectionChanged(QItemSelection(), QItemSelection()); // enable / disable buttons
@@ -179,8 +191,9 @@ TemplateWidget::TemplateWidget(Map* map, MapView* main_view, MapEditorController
 	connect(adjust_button, SIGNAL(clicked(bool)), this, SLOT(adjustClicked(bool)));
 	//connect(group_button, SIGNAL(clicked(bool)), this, SLOT(groupClicked()));
 	connect(position_button, SIGNAL(clicked(bool)), this, SLOT(positionClicked(bool)));
-	connect(more_button_menu, SIGNAL(triggered(QAction*)), this, SLOT(moreActionClicked(QAction*)));
+	//connect(more_button_menu, SIGNAL(triggered(QAction*)), this, SLOT(moreActionClicked(QAction*)));
 	
+	connect(map, SIGNAL(templateAdded(int,Template*)), this, SLOT(templateAdded(int,Template*)));
 	connect(controller, SIGNAL(templatePositionDockWidgetClosed(Template*)), this, SLOT(templatePositionDockWidgetClosed(Template*)));
 }
 TemplateWidget::~TemplateWidget()
@@ -194,11 +207,11 @@ TemplateWidget::~TemplateWidget()
 
 void TemplateWidget::addTemplateAt(Template* new_template, int pos)
 {
-	int row;
+	/*int row;
 	if (pos >= 0)
 		row = template_table->rowCount() - 1 - ((pos >= map->getFirstFrontTemplate()) ? (pos + 1) : pos);
 	else
-		row = template_table->rowCount() - 1 - map->getFirstFrontTemplate();
+		row = template_table->rowCount() - 1 - map->getFirstFrontTemplate();*/
 	
 	if (pos < map->getFirstFrontTemplate())
 		map->setFirstFrontTemplate(map->getFirstFrontTemplate() + 1);
@@ -206,18 +219,10 @@ void TemplateWidget::addTemplateAt(Template* new_template, int pos)
 		pos = map->getFirstFrontTemplate() - 1;
 	
 	// Add template and make it visible in the currently active view; TODO: currently, it is made visible in the main view -> support multiple views
-	map->addTemplate(new_template, pos);
-	TemplateVisibility* vis = main_view->getTemplateVisibility(new_template);
-	vis->visible = true;
-	vis->opacity = 1;
+	map->addTemplate(new_template, pos, main_view);
 	map->setTemplateAreaDirty(pos);
 	
-	++row;	// always insert new row below the selected row
-	template_table->insertRow(row);
-	addRow(row);
-	template_table->setCurrentCell(row, 3);
-	
-	map->setTemplatesDirty();	// TODO: redraw map widget(s)
+	map->setTemplatesDirty();
 }
 Template* TemplateWidget::showOpenTemplateDialog(QWidget* dialog_parent, MapView* main_view)
 {
@@ -238,17 +243,30 @@ Template* TemplateWidget::showOpenTemplateDialog(QWidget* dialog_parent, MapView
 		return NULL;
 	}
 	
-	if (!new_temp->open(dialog_parent, main_view))
+	if (!new_temp->preLoadConfiguration(dialog_parent))
 	{
 		delete new_temp;
 		return NULL;
 	}
 	
-	if (!new_temp->isTemplateValid())
+	if (!new_temp->loadTemplateFile(true))
 	{
 		QMessageBox::warning(dialog_parent, tr("Error"), tr("Cannot open template:\n%1\n\nFailed to load template. Does the file exist and is it valid?").arg(path));
 		delete new_temp;
 		return NULL;
+	}
+	
+	if (!new_temp->postLoadConfiguration(dialog_parent))
+	{
+		delete new_temp;
+		return NULL;
+	}
+	
+	// If the template is not georeferenced, position it at the viewport midpoint
+	if (!new_temp->isTemplateGeoreferenced())
+	{
+		new_temp->setTemplateX(main_view->getPositionX());
+		new_temp->setTemplateY(main_view->getPositionY());
 	}
 	
 	return new_temp;
@@ -324,12 +342,17 @@ void TemplateWidget::deleteTemplate()
 	assert(pos >= 0);
 	
 	map->setTemplateAreaDirty(pos);
-	map->deleteTemplate(pos);
+	
+	if (Settings::getInstance().getSettingCached(Settings::Templates_KeepSettingsOfClosed).toBool())
+		map->closeTemplate(pos);
+	else
+		map->deleteTemplate(pos);
+	
 	template_table->removeRow(template_table->currentRow());
 	if (pos < map->getFirstFrontTemplate())
 		map->setFirstFrontTemplate(map->getFirstFrontTemplate() - 1);
 	
-	map->setTemplatesDirty();	// TODO: redraw map widget(s)
+	map->setTemplatesDirty();
 }
 void TemplateWidget::duplicateTemplate()
 {
@@ -559,16 +582,51 @@ void TemplateWidget::selectionChanged(const QItemSelection& selected, const QIte
 	move_up_button->setEnabled(current_row >= 1 && !multiple_rows_selected);
 	move_down_button->setEnabled(current_row < template_table->rowCount() - 1 && current_row != -1 && !multiple_rows_selected);
 	
-	bool enable_active_buttons = current_row >= 0 && !map_row;
+	QString active_group_title;
+	QString georef_text;
+	bool georef_active;
+	if (current_row < 0)
+	{
+		active_group_title = tr("No selection");
+		georef_text = "-";
+		georef_active = false;
+	}
+	else if (multiple_rows_selected)
+	{
+		active_group_title = tr("Multiple templates selected");
+		georef_text = "-";
+		georef_active = false;
+	}
+	else if (map_row)
+	{
+		active_group_title = tr("- Map -");
+		georef_active = !map->getGeoreferencing().isLocal();
+		georef_text = georef_active ? tr("yes") : tr("no");
+	}
+	else
+	{
+		active_group_title = temp->getTemplateFilename();
+		georef_active = temp->isTemplateGeoreferenced();
+		georef_text = georef_active ? tr("yes") : tr("no");
+	}
+	active_buttons_group->setTitle(active_group_title);
+	georef_button->setText(tr("Georeferenced: %1").arg(georef_text));
+	georef_button->setChecked(georef_active);
+	
+	bool enable_active_buttons = current_row >= 0 && !map_row && temp && !multiple_rows_selected;
 	active_buttons_group->setEnabled(enable_active_buttons);
 	if (enable_active_buttons)
 	{
-		move_by_hand_button->setEnabled(!multiple_rows_selected);
-		adjust_button->setEnabled(!multiple_rows_selected);
+		// TODO: enable changing georeferencing state
+		georef_button->setEnabled(false);
+		
+		move_by_hand_button->setEnabled(!temp->isTemplateGeoreferenced());
+		adjust_button->setEnabled(!temp->isTemplateGeoreferenced());
+		position_button->setEnabled(!temp->isTemplateGeoreferenced());
+		
 		// TODO: Implement and enable buttons again
 		//group_button->setEnabled(false); //multiple_rows_selected || (!multiple_rows_selected && map->getTemplate(posFromRow(current_row))->getTemplateGroup() >= 0));
-		position_button->setEnabled(!multiple_rows_selected);
-		more_button->setEnabled(false); // !multiple_rows_selected);
+		//more_button->setEnabled(false); // !multiple_rows_selected);
 	}
 	
 	if (multiple_rows_selected)
@@ -591,7 +649,7 @@ void TemplateWidget::currentCellChange(int current_row, int current_column, int 
         if (!temp)
             return;
 
-        if (!temp->isTemplateValid())
+        if (temp->getTemplateState() == Template::Invalid)
             changeTemplateFile(current_row);
     }
 }
@@ -606,6 +664,12 @@ void TemplateWidget::cellDoubleClick(int row, int column)
 		
 		changeTemplateFile(row);
 	}
+}
+
+void TemplateWidget::updateDeleteButtonText()
+{
+	bool keep_transformation_of_closed_templates = Settings::getInstance().getSettingCached(Settings::Templates_KeepSettingsOfClosed).toBool();
+	delete_button->setText(keep_transformation_of_closed_templates ? tr("Close") : tr("Delete"));
 }
 
 void TemplateWidget::moveByHandClicked(bool checked)
@@ -656,6 +720,14 @@ void TemplateWidget::positionClicked(bool checked)
 void TemplateWidget::moreActionClicked(QAction* action)
 {
 	// TODO
+}
+
+void TemplateWidget::templateAdded(int pos, Template* temp)
+{
+	int row = rowFromPos(pos);
+	template_table->insertRow(row);
+	addRow(row);
+	template_table->setCurrentCell(row, 3);
 }
 
 void TemplateWidget::templatePositionDockWidgetClosed(Template* temp)
@@ -732,7 +804,7 @@ void TemplateWidget::updateRow(int row)
         vis = main_view->getTemplateVisibility(temp);
         group = temp->getTemplateGroup();
         name = temp->getTemplateFilename();
-        valid = temp->isTemplateValid();
+        valid = temp->getTemplateState() != Template::Invalid;
     }
     else
     {
@@ -750,6 +822,7 @@ void TemplateWidget::updateRow(int row)
 
 	react_to_changes = true;
 }
+
 int TemplateWidget::posFromRow(int row)
 {
 	int pos = template_table->rowCount() - 1 - row;
@@ -762,6 +835,13 @@ int TemplateWidget::posFromRow(int row)
 	else
 		return pos;
 }
+
+int TemplateWidget::rowFromPos(int pos)
+{
+	assert(pos >= 0);
+	return map->getNumTemplates() - 1 - ((pos >= map->getFirstFrontTemplate()) ? pos : (pos - 1));
+}
+
 Template* TemplateWidget::getCurrentTemplate()
 {
 	int current_row = template_table->currentRow();
@@ -775,17 +855,14 @@ Template* TemplateWidget::getCurrentTemplate()
 
 void TemplateWidget::changeTemplateFile(int row)
 {
-	QString path = QFileDialog::getOpenFileName(this, tr("Find the moved template file"), QString(), tr("All files (*.*)"));
-	path = QFileInfo(path).canonicalFilePath();
-	if (path.isEmpty())
-		return;
-	
 	int pos = posFromRow(row);
 	Template* temp = (row >= 0 && pos >= 0) ? map->getTemplate(pos) : NULL;
 	assert(temp);
 	
-	if (temp->changeTemplateFile(path))
+	if (temp->execSwitchTemplateFileDialog(this))
+	{
 		updateRow(row);
-	else
-		QMessageBox::warning(this, tr("Error"), tr("Cannot change the template to this file! Is the format of the file correct for this template type?"));
+		temp->setTemplateAreaDirty();
+		map->setTemplatesDirty();
+	}
 }
